Remove the certificate checking defaults used by LWP::UserAgent
When using this tool through proxychains to then proxy through an upstream content inspection proxy such as BurpSuite, the tool fails to validate the certificate and as a result does not work.  As this tool is not designed for user security, the strict certificate checking defaults seem irrelevant and so have been disabled.